+++ src/variables.mjs
@@ -23,7 +23,7 @@ export { default as doubleval } from './variables/doubleval.mjs';
 export { default as empty } from './variables/empty.mjs';
 export { default as get_debug_type } from './variables/get_debug_type.mjs';
 export { default as gettype } from './variables/gettype.mjs';
-export { default as intval } from './variables/intval.mjs';
+export { default as to_int } from './variables/to_int.mjs';
 export { default as isset } from './variables/isset.mjs';
 export { default as strval } from './variables/strval.mjs';
 export { default as print_r } from './variables/print_r.mjs';