+++ src/variables/settype.mjs
@@ -1,7 +1,7 @@
 import arrval from './to_array.mjs';
 import boolval from './to_bool.mjs';
 import floatval from './floatval.mjs';
-import intval from './intval.mjs';
+import intval from './to_int.mjs';
 import objval from './objval.mjs';
 import strval from './strval.mjs';