+++ tests/variables.global.test.mjs
@@ -27,7 +27,7 @@ test('variables.global', () => {
     expect(empty).not.toBeUndefined();
     expect(get_debug_type).not.toBeUndefined();
     expect(gettype).not.toBeUndefined();
-    expect(intval).not.toBeUndefined();
+    expect(to_int).not.toBeUndefined();
     expect(isset).not.toBeUndefined();
     expect(strval).not.toBeUndefined();
     expect(print_r).not.toBeUndefined();