+++ tests/variables.test.mjs
@@ -7,7 +7,7 @@ import {
     floatval,
     get_debug_type,
     gettype,
-    intval,
+    to_int,
     is_array,
     is_bool,
     is_callable,
@@ -66,7 +66,7 @@ test('variables', () => {
     expect(empty).not.toBeUndefined();
     expect(get_debug_type).not.toBeUndefined();
     expect(gettype).not.toBeUndefined();
-    expect(intval).not.toBeUndefined();
+    expect(to_int).not.toBeUndefined();
     expect(isset).not.toBeUndefined();
     expect(strval).not.toBeUndefined();
     expect(print_r).not.toBeUndefined();